rename: FE폴더구조 추가변경
